### 2024-09-16 #7141 @csoneson
👋🏻 @MicroPhen - I will handle your paper, and will start by looking for a couple of suitable reviewers. Before that, however, I would like to ask you to significantly cut down on the length of your paper - it's currently almost 6,000 words, and the target is 250-1,000 (see the [JOSS guidelines](https://joss.readthedocs.io/en/latest/paper.html) for more details). Once this is done, please let me know here (you can generate a new proof by typing `@editorialbot generate pdf` in a comment in this issue). Thanks!
