### 2024-09-24 #7141 @Y0dler
Hello @csoneson, thanks for handling our paper.

Regarding the page number, I must admit we apparently overlooked that part. The problem is just that we realized during an earlier attempt to publish this content that a) most people in our primary target audience of biologists and chemists, who usually deal with chromatographic peak data, have very little to no experience with Bayesian statistics and b) we have to explain the model structure in order to set ourselves apart from other instances where Bayesian statistics were employed in one way or another to peak recognition or related topics. We will do our best to shorten the paper and shift some sections from the manuscript to the documentation but it would be very difficult to arrive at < 1000 words and still address the initially mentioned problems.
Is this an immovable upper limit or is it just the case that 6000 words are too far off from the target?
