### 2024-09-25 #7141 @csoneson
Hi @Y0dler - 6,000 words is definitely too far from the target, we can accept a bit over 1,000 words but I think the goal should be to get reasonably close. As mentioned in the guidelines, JOSS papers are not intended to be "regular" research papers, and longer descriptions of models etc may fit better in the documentation. Thanks!
