### 2024-09-25 3ee22eb
reduce paper length; oursource model structure to .md file
