### 2024-09-25 4dd60d4
trim down validation section
