### 2024-09-26 ea3c0ac
Remove validation completely
