+++ paper/paper.md
@@ -68,26 +68,6 @@ Bayesian inference is conducted utilizing the PyMC package [@RN150] with the ext
 Both model selection and analysis of inference data objects are realized with the ArviZ package [@RN147].
 Since the inference data is stored alongside graphs and report sheets, users may employ the ArviZ package or others for further analysis of the results if necessary.
 
-## Validation of $\texttt{PeakPerformance}$
-Several stages of validation were employed to prove the suitability of $\texttt{PeakPerformance}$ for chromatographic peak data analysis.
-For the first and second tests, 500 random data sets were generated with the NumPy random module [@harris2020array] by drawing from the normal distributions detailed in Table 1 except for the mean parameter which was held constant at a value of 6.
-Additionally, in case of the second test, the area was set to 8 and the skewness parameter $\alpha$ to 1.
-Subsequently, normally distributed random noise ($\mathcal{N}(0, 0.6)$ or $\mathcal{N}(0, 1.2)$ for data sets with the tag "higher noise") was added to each data point.
-The amount of data points per time was chosen based on an LC-MS/MS method routinely utilized by the authors and accordingly set to one data point per 1.8 s.
-
-__Table 1:__ Normal distributions from which parameters were drawn randomly to create synthetic data sets for the validation of $\texttt{PeakPerformance}$.
-
-| **parameter**      | **model (1st test)**    | **model (2nd test)**  |
-| ------------------ | ----------------------- | ----------------------- |
-| area               | $\mathcal{N}(8, 0.5)$   | -                       |
-| standard deviation | $\mathcal{N}(0.5, 0.1)$ | $\mathcal{N}(0.5, 0.1)$ |
-| skewness           | $\mathcal{N}(0, 2)$    | -                       |
-| baseline intercept | $\mathcal{N}(25, 1)$    | $\mathcal{N}(25, 1)$    |
-| baseline slope     | $\mathcal{N}(0, 1)$     | $\mathcal{N}(0, 1)$     |
-
-For the third and final test, experimental peak data was analyzed with both $\texttt{PeakPerformance}$ (version 0.7.0) and Sciex MultiQuant (version 3.0.3) with human supervision, i.e. the results were visually inspected and corrected if necessary.
-The data set consisted of 192 signals comprised of 123 single peaks, 50 peaks as part of double peaks, and 19 noise signals.
-
 
 # Results and Discussion
 
@@ -138,45 +118,9 @@ __Table 2:__ Depiction of the results for the most important peak parameters of 
 ![](./summary_joint.svg){width="100%"}
 
 In this case, the fits were successful and convergence was reached for all parameters.
-Most notably and for the first time, the measurement noise was taken into account when determining the peak area as represented by its standard deviation and as can be observed in the posterior predictive plots where the noisy data points fall within the boundary of the 95 % HDI.
+Most notably and for the first time, the measurement noise was taken into account when determining the peak area as represented by its standard deviation and as can be observed in the posterior predictive plots where the noisy data points fall within the boundary of the 95 % HDI.\\
+In the documentation, there is a study featuring simulated and experimental data to validate $\texttt{PeakPerformance}$'s results against a commercially available vendor software for peak integration showing that comparable results are indeed obtained.
 
-
-## Validation
-In the first stage of validation, peak fitting with normal and skew normal peak models was tested regarding the ability to reproduce the ground truth of randomly generated noisy synthetic data sets.
-The arithmetic means portrayed in Figure 3a were calculated based on a measure of similarity
-
-$$\tag{12}F_{y / \hat{y}} = \frac{y}{\hat{y}}$$
-
-where $y$ represents the estimated parameter value and $\hat{y}$ its pertaining ground truth.
[…]
