### 2024-09-26 #7141 @Y0dler
Hello again, @csoneson, I removed many sections and trimmed the remaining ones down to arrive at 1693 words (counting from the beginning of `summary` to the end of `conclusion`).  Is this acceptable?
