### 2024-09-26 6a4bd65
Remove peak model composition file for now
