+++ docs/source/markdown/PeakPerformance_validation.md
@@ -1,0 +1,67 @@
+# Validation of $\texttt{PeakPerformance}$
+
+## Materials and Methods
+Several stages of validation were employed to prove the suitability of $\texttt{PeakPerformance}$ for chromatographic peak data analysis.
+The goals were to showcase the efficacy of $\texttt{PeakPerformance}$ utilizing noisy synthetic data, to investigate cases where a peak could reasonably be fit with either of the single peak models, and finally to use experimental data to compare results obtained with $\texttt{PeakPerformance}$ to those from the commercial vendor software Sciex MultiQuant.
+
+For the first test, 500 random data sets were generated with the NumPy random module [@harris2020array] by drawing from the normal distributions detailed in Table 1 except for the mean parameter which was held constant at a value of 6.
+Subsequently, normally distributed random noise ($\mathcal{N}(0, 0.6)$ or $\mathcal{N}(0, 1.2)$ for data sets with the tag "higher noise") was added to each data point.
+The amount of data points per time was chosen based on an LC-MS/MS method routinely utilized by the authors and accordingly set to one data point per 1.8 s.
+
+__Table 1:__ Normal distributions from which parameters were drawn randomly to create synthetic data sets for the validation of $\texttt{PeakPerformance}$.
+
+| **parameter**      | **model (1st test)**    | **model (2nd test)**  |
+| ------------------ | ----------------------- | ----------------------- |
+| area               | $\mathcal{N}(8, 0.5)$   | -                       |
+| standard deviation | $\mathcal{N}(0.5, 0.1)$ | $\mathcal{N}(0.5, 0.1)$ |
+| skewness           | $\mathcal{N}(0, 2)$    | -                       |
+| baseline intercept | $\mathcal{N}(25, 1)$    | $\mathcal{N}(25, 1)$    |
+| baseline slope     | $\mathcal{N}(0, 1)$     | $\mathcal{N}(0, 1)$     |
+
+In marginal cases when the shape of a single peak had a slight skew, the automated model selection would at times settle on a normal or a skew normal model.
+Therefore, it was relevant to investigate whether this choice would lead to a significant discrepancy in estimated peak parameters.
+Accordingly, for the second test synthetic data sets were generated with the NumPy random module according to Table 1 and noise was added as described before.
+The residual parameters were held constant, i.e. the mean was fixed to 6, the area to 8, and the skewness parameter $\alpha$ to 1.
+
+For the third and final test, experimental peak data was analyzed with both $\texttt{PeakPerformance}$ (version 0.7.0) and Sciex MultiQuant (version 3.0.3) with human supervision, i.e. the results were visually inspected and corrected if necessary.
+The data set consisted of 192 signals comprised of 123 single peaks, 50 peaks as part of double peaks, and 19 noise signals.
+
+
+## Results and Discussion
+In the first stage of validation, peak fitting with normal and skew normal peak models was tested regarding the ability to reproduce the ground truth of randomly generated noisy synthetic data sets.
+The arithmetic means portrayed in Figure 1a were calculated based on a measure of similarity
+
+$$\tag{12}F_{y / \hat{y}} = \frac{y}{\hat{y}}$$
+
+where $y$ represents the estimated parameter value and $\hat{y}$ its pertaining ground truth.
+As they exhibit values close to 1, this demonstrates a near identity between estimation and ground truth.
+Additionally, the normal-shaped peak model was paired with skew normally distributed noisy data and vice versa.
+In both cases, $\sigma$ was not reproduced well, especially by the normal-shaped model.
+Nevertheless, the peak area and height were still identified correctly with the skew normal model and merely slightly underestimated by the normal model.
+
+![](./Fig6_PP-validation.png)
[…]
