+++ docs/source/index.rst
@@ -37,6 +37,8 @@ The documentation features various notebooks that demonstrate the usage and inve
    markdown/Installation
    markdown/Peak_model_composition
    markdown/PeakPerformance_validation
+   markdown/PeakPerformance_workflow
+   markdown/Diagnostic_plots
    markdown/How_to_adapt_PeakPerformance_to_your_data
    notebooks/Ex1_Simple_Pipeline.ipynb
    notebooks/Ex2_Custom_Use_of_PeakPerformance.ipynb
