### 2024-09-26 ed2bdb4
add images etc. to markdown folder
