+++ docs/source/markdown/literature.bib
@@ -1,0 +1,213 @@
+@softmisc{nutpie,
+  author   = {Seyboldt, Adrian and {PyMC Developers}},
+  keywords = {Software},
+  license  = {MIT},
+  title    = {{nutpie}},
+  url      = {https://github.com/pymc-devs/nutpie}
+}
+
+@article{scipy,
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
+             Haberland, Matt and Reddy, Tyler and Cournapeau, David and
+             Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
+             Bright, Jonathan and {van der Walt}, St{\'e}fan J. and
+             Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and
+             Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and
+             Kern, Robert and Larson, Eric and Carey, C J and
+             Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and
+             {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and
+             Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and
+             Harris, Charles R. and Archibald, Anne M. and
+             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
+             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
+  title   = {{{SciPy} 1.0: {F}undamental Algorithms for Scientific Computing in {P}ython}},
+  journal = {Nature Methods},
+  year    = {2020},
+  volume  = {17},
+  pages   = {261--272},
+  adsurl  = {https://rdcu.be/b08Wh},
+  doi     = {10.1038/s41592-019-0686-2}
+}
+
+@article{matplotlib,
+  author    = {Hunter, J. D.},
+  title     = {Matplotlib: A 2D graphics environment},
+  journal   = {Computing in Science \& Engineering},
+  volume    = {9},
+  number    = {3},
+  pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+               application development, interactive scripting, and publication-quality
+               image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
+
+@softmisc{matplotlibzenodo,
+  author    = {{The Matplotlib Development Team}},
+  title     = {Matplotlib: Visualization with Python},
+  keywords  = {software},
+  month     = may,
+  year      = 2024,
+  publisher = {Zenodo},
+  version   = {v3.9.0},
+  doi       = {10.5281/zenodo.11201097},
+  url       = {https://doi.org/10.5281/zenodo.11201097}
+}
+
+@article{RN173,
+  author  = {Hoffmann, Matthew D. and Gelman, Andrew},
+  title   = {The No-U-Turn Sampler: Adaptively Setting Path Lengths in Hamiltonian Monte Carlo},
+  journal = {Journal of Machine Learning Research},
+  volume  = {15},
+  year    = {2014},
+  type    = {Journal Article}
+}
+
+@article{RN150,
+  author  = {Abril-Pla, O. and Andreani, V. and Carroll, C. and Dong, L. and Fonnesbeck, C. J. and Kochurov, M. and Kumar, R. and Lao, J. and Luhmann, C. C. and Martin, O. A. and Osthege, M. and Vieira, R. and Wiecki, T. and Zinkov, R.},
[…]
