### 2024-09-28 #7141 @csoneson
👋🏻 @Armavica, @adafede, @bittremieux - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> PeakPerformance - A tool for Bayesian inference-based fitting of LC-MS/MS peaks
https://github.com/openjournals/joss-reviews/issues/7141

The checklist-based review is carried out on GitHub, more details can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Thanks in advance!
