### 2024-09-29 2ee7b10
Enable MyST extensions to fix math rendering
