+++ docs/source/markdown/Peak_model_composition.md
@@ -9,13 +9,17 @@ For example, when only positive values were acceptable or when 0 was not a permi
 
 Regarding shared model elements across all intensity functions, one such component of all models presented hereafter is the likelihood function
 
-$$\tag{1}L \sim \mathcal{N}(y, \mathrm{noise})$$
+$$
+L \sim \mathcal{N}(y, \mathrm{noise})
+$$ (eqn:likelihood)
 
 with $y$ as the predicted intensity and $\mathrm{noise}$ as the free parameter describing the standard deviation of measurement noise.
 This definition encodes the assumption that observed intensities are the result of normally distributed noise around the true intensity values of a peak.
 In turn, the noise parameter is defined as
 
-$$\tag{2}\mathrm{noise} \sim \mathrm{LogNormal}(\log_{10} \mathrm{max}(10, \mathrm{noise}_{\mathrm{guess}}), 1)$$
+$$
+\mathrm{noise} \sim \mathrm{LogNormal}(\log_{10} \mathrm{max}(10, \mathrm{noise}_{\mathrm{guess}}), 1)
+$$ (eqn:noise)
 
 The log-normal distribution where the logarithm of the random variable follows a normal distribution was chosen partly to exclude negative values from the solution space and also due to its shape attributing a higher fraction of the probability mass to lower values provided the standard deviation is defined sufficiently high.
 This prior is defined in a raw data-dependent manner as the $\mathrm{noise}_{\mathrm{guess}}$ amounts to the standard deviation of the differences of the first and final 15 % of intensity values included in a given time frame and their respective mean values.
@@ -23,28 +27,30 @@ This prior is defined in a raw data-dependent manner as the $\mathrm{noise}_{\ma
 The intensity function itself is defined as the sum of a linear baseline function and a peak intensity function, the latter of which is composed of a given distribution's probability density function (PDF) scaled up to the peak size by the area or height parameter.
 The linear baseline
 
-$$\tag{3}y_{\mathrm{baseline}}(t) = at+b$$
+$$
+y_{\mathrm{baseline}}(t) = at+b
+$$ (eqn:baseline)
 
 features the slope and intersect parameters $a$ and $b$, respectively, both of which were assigned a normally distributed prior.
 The data-dependent guesses for these priors are obtained by constructing a line through the means of the first and last three data points of a given intensity data set which oftentimes already resulted in a good fit.
 Hence, the determined values for slope ($a_{\mathrm{guess}}$) and intercept ($b_{\mathrm{guess}}$) are used as the means for their pertaining priors and the standard deviations are defined as fractions of them with minima set to 0.5 and 0.05, respectively.
 Here, the exact definition of the standard deviations was less important than simply obtaining an uninformative prior which, while based on the rough fit for the baseline, possesses a sufficient degree of independence from it, thus allowing deviations by the Bayesian parameter estimation.
 
-$$\tag{4}
+$$
     a \sim
     \begin{cases}
         \mathcal{N}(a_{\mathrm{guess}}, \frac{|a_{\mathrm{guess}}|}{5}) & \mathrm{if}\ \frac{|a_{guess}|}{5}\geq0.5\\
         \mathcal{N}(a_{\mathrm{guess}}, 0.5) & \mathrm{otherwise}\\
     \end{cases}
-$$
+$$ (eqn:guess_a)
 
-$$\tag{5}
+$$
     b \sim
     \begin{cases}
[…]
