+++ docs/source/markdown/PeakPerformance_workflow.md
@@ -4,10 +4,10 @@ bibliography:
 ---
 
 # PeakPerformance workflow
-$\texttt{PeakPerformance}$ accommodates the use of a pre-manufactured data pipeline for standard applications as well as the creation of custom data pipelines using only its core functions.
+`PeakPerformance` accommodates the use of a pre-manufactured data pipeline for standard applications as well as the creation of custom data pipelines using only its core functions.
 The provided data analysis pipeline was designed in a user-friendly way and requires minimal programming knowledge ([Fig. 1](#fig_w1)).
 As portrayed in an example notebook in the code repository, only a few simple Python commands need to be executed.
-Instead of relying on these convenience functions, experienced users can also directly access the core functions of $\texttt{PeakPerformance}$ for a more flexible application which is demonstrated in yet another example notebook.
+Instead of relying on these convenience functions, experienced users can also directly access the core functions of `PeakPerformance` for a more flexible application which is demonstrated in yet another example notebook.
 
 ```{figure-md} fig_w1
 ![](./Fig3_PP-standalone.png)
@@ -15,14 +15,14 @@ Instead of relying on these convenience functions, experienced users can also di
 __Figure 1:__ Overview of the pre-manufactured data analysis pipeline featured in `PeakPerformance`.
 ```
 
-Before using $\texttt{PeakPerformance}$, the user has to supply raw data files containing a NumPy array with time in the first and intensity in the second dimension.
-For each peak, such a file has to be provided according to the naming convention specified in $\texttt{PeakPerformance}$'s documentation and gathered in one directory.
+Before using `PeakPerformance`, the user has to supply raw data files containing a NumPy array with time in the first and intensity in the second dimension.
+For each peak, such a file has to be provided according to the naming convention specified in `PeakPerformance`'s documentation and gathered in one directory.
 If a complete time series of a 30 - 90 min LC-MS/MS run were to be submitted to the program, however, the target peak would make up an extremely small portion of this data.
 Additionally, other peaks with the same mass and fragmentation pattern may have been observed at different retention times.
 Therefore, it was decided from the outset that in order to enable proper peak fitting, only a fraction of such a time series with a range of 3 - 5 times the peak width and roughly centered on the target peak would be accepted as an input.
 This guarantees that there is a sufficient number of data points at the beginning and end of the time frame for estimating the baseline and noise level, as well.
-The provided data pipeline starts by defining a path to this raw data directory and one to a local clone of the $\texttt{PeakPerformance}$ code repository.
-Using the $\texttt{prepare\_model\_selection()}$ method, an Excel template file ("Template.xlsx") for inputting user information is prepared and stored in the raw data directory.
+The provided data pipeline starts by defining a path to this raw data directory and one to a local clone of the `PeakPerformance` code repository.
+Using the `prepare_model_selection()` function, an Excel template file ("Template.xlsx") for inputting user information is prepared and stored in the raw data directory.
 It is the user's task, then, to select the settings for the pipeline within the file itself.
 Accordingly, the file contains detailed explanations of all parameters and the parsing functions of the software feature clear error messages in case mandatory entries are missing or filled out incorrectly.
 
@@ -33,21 +33,21 @@ It is then assumed that within one batch run, all instances of a mass trace acro
 For this purpose, the user must provide the name of an acquisition, i.e. sample, where a clear and representative peak for the given mass trace was observed.
 If e.g. a standard mixture containing all targets was measured, this would be considered a prime candidate.
 An additional feature lets the user exclude specific model types to save computation time and improve the accuracy of model selection by for example excluding double peak models when a single peak was observed.
[…]
