### 2024-09-29 d642265
Fix citations and bibliography
