+++ docs/source/conf.py
@@ -37,6 +37,7 @@ extensions = [
     "numpydoc",
     "myst_nb",
     "sphinx_book_theme",
+    "sphinxcontrib.bibtex",
     "sphinxcontrib.mermaid",
 ]
 myst_enable_extensions = [
@@ -48,6 +49,9 @@ myst_enable_extensions = [
     "tasklist",
 ]
 nb_execution_mode = "off"
+bibtex_bibfiles = ["literature.bib"]
+bibtex_default_style = "plain"
+bibtex_reference_style = "label"
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
