+++ docs/source/markdown/PeakPerformance_workflow.md
@@ -1,8 +1,3 @@
----
-bibliography:
-  - literature.bib
----
-
 # PeakPerformance workflow
 `PeakPerformance` accommodates the use of a pre-manufactured data pipeline for standard applications as well as the creation of custom data pipelines using only its core functions.
 The provided data analysis pipeline was designed in a user-friendly way and requires minimal programming knowledge ([Fig. 1](#fig_w1)).
@@ -35,7 +30,7 @@ If e.g. a standard mixture containing all targets was measured, this would be co
 An additional feature lets the user exclude specific model types to save computation time and improve the accuracy of model selection by for example excluding double peak models when a single peak was observed.
 Upon provision of the required information, the automated model selection can be started using the `model_selection()` function from the pipeline module and will be performed successively for each mass trace.
 Essentially, every type of model which has not been excluded by the user needs to be instantiated, sampled, and the log-likelihood needs to be calculated.
-Subsequently, the results for each model are ranked with the `compare()` function of the ArviZ package based on Pareto-smoothed importance sampling leave-one-out cross-validation (LOO-PIT) [@RN146; @RN145].
+Subsequently, the results for each model are ranked with the `compare()` function of the ArviZ package based on Pareto-smoothed importance sampling leave-one-out cross-validation (LOO-PIT) {cite}`RN146,RN145`.
 This function returns a DataFrame showing the results of the models in order of their placement on the ranking which is decided by the expected log pointwise predictive density.
 The best model for each mass trace is then written to the Excel template file.
 
@@ -44,16 +39,16 @@ The first step consists of parsing the information from the Excel sheet.
 Since the pipeline, just like model selection, acts successively, a time series is read from its data file next and the information contained in the name of the file according to the naming convention is parsed.
 All this information is combined in an instance of `PeakPerformance`'s `UserInput` class acting as a centralized source of data for the program.
 Depending on whether the "pre-filtering" option was selected, an optional filtering step will be executed to reject signals where clearly no peak is present before sampling, thus saving computation time.
-This filtering step uses the `find_peaks()` function from the SciPy package [@scipy] which simply checks for data points directly neighboured by points with lower intensity values.
+This filtering step uses the `find_peaks()` function from the SciPy package {cite}`scipy` which simply checks for data points directly neighboured by points with lower intensity values.
 If no data points within a certain range around the expected retention time of an analyte fulfill this most basic requirement of a peak, the signal is rejected.
 Furthermore, if none of the candidate data points exceed a signal-to-noise ratio threshold defined by the user, the signal will also be discarded.
 Depending on the origin of the samples, this step may reject a great many signals before sampling saving potentially hours of computation time across a batch run of the `PeakPerformance` pipeline.
 For instance, in bioreactor cultivations, a product might be quantified but if it is only produced during the stationary growth phase, it will not show up in early samples.
 Another pertinent example of such a use case are isotopic labeling experiments for which every theoretically achievable mass isotopomer needs to be investigated, yet depending on the input labeling mixture, the majority of them might not be present in actuality.
-Upon passing the first filter, a Markov chain Monte Carlo (MCMC) simulation is conducted using a No-U-Turn Sampler (NUTS) [@RN173], preferably - if installed in the Python environment - the nutpie sampler [@nutpie] due to its highly increased performance compared to the default sampler of PyMC.
+Upon passing the first filter, a Markov chain Monte Carlo (MCMC) simulation is conducted using a No-U-Turn Sampler (NUTS) {cite}`RN173`, preferably - if installed in the Python environment - the nutpie sampler {cite}`nutpie` due to its highly increased performance compared to the default sampler of PyMC.
[…]
