+++ docs/source/markdown/PeakPerformance_validation.md
@@ -4,7 +4,7 @@
 Several stages of validation were employed to prove the suitability of `PeakPerformance` for chromatographic peak data analysis.
 The goals were to showcase the efficacy of `PeakPerformance` utilizing noisy synthetic data, to investigate cases where a peak could reasonably be fit with either of the single peak models, and finally to use experimental data to compare results obtained with `PeakPerformance` to those from the commercial vendor software Sciex MultiQuant.
 
-For the first test, 500 random data sets were generated with the NumPy random module [@harris2020array] by drawing from the normal distributions detailed in Table 1 except for the mean parameter which was held constant at a value of 6.
+For the first test, 500 random data sets were generated with the NumPy random module {cite}`harris2020array` by drawing from the normal distributions detailed in Table 1 except for the mean parameter which was held constant at a value of 6.
 Subsequently, normally distributed random noise ($\mathcal{N}(0, 0.6)$ or $\mathcal{N}(0, 1.2)$ for data sets with the tag "higher noise") was added to each data point.
 The amount of data points per time was chosen based on an LC-MS/MS method routinely utilized by the authors and accordingly set to one data point per 1.8 s.
 
@@ -75,3 +75,6 @@ By showing not only the mean area ratio of all peaks but also the ones for the s
 In case of this data set, two low quality double peaks in particular inflated the variance significantly which may not be representative for other data sets.
 It has to be stated, too, that the prevalence of manual re-integration of double peaks in MQ might have introduced a user-specific bias, thereby increasing the final variance.
 Nevertheless, it could be shown that `PeakPerformance` yields comparable peak area results to a commercially available vendor software.
+
+```{bibliography}
+```
