+++ docs/source/literature.bib
@@ -1,4 +1,4 @@
-@softmisc{nutpie,
+@misc{nutpie,
   author   = {Seyboldt, Adrian and {PyMC Developers}},
   keywords = {Software},
   license  = {MIT},
@@ -44,7 +44,7 @@
   year      = 2007
 }
 
-@softmisc{matplotlibzenodo,
+@misc{matplotlibzenodo,
   author    = {{The Matplotlib Development Team}},
   title     = {Matplotlib: Visualization with Python},
   keywords  = {software},
@@ -83,6 +83,7 @@
   author  = {Kruschke, John K.},
   title   = {Doing Bayesian Data Analysis},
   edition = {1st Edition},
+  publisher={Academic Press},
   isbn    = {9780123814852},
   year    = {2010},
   type    = {Book}
