### 2024-09-29 9307b88
Format installation page with admonitions
