### 2024-10-03 #7313 @editorialbot
[REVIEW]: PeakPerformance - A tool for Bayesian inference-based fitting of LC-MS/MS peaks
**Submitting author:** @MicroPhen (<a href="http://orcid.org/0000-0001-9784-3626">Stephan Noack</a>)
**Repository:** https://github.com/JuBiotech/peak-performance/
**Branch with paper.md** (empty if default branch): main
**Version:** v0.7.2
