### 2024-10-04 2bdbe56
add raw data preparation to documentation
