### 2024-10-04 #7313 @Y0dler
First of all: thank you very much for agreeing to review our manuscript and program.

I just wanted to inform you that I created a branch to improve and streamline our GitHub repo's landing page a moment ago as it is a little bit outdated and confusing. More importantly, though, here is the link to our [documentation](https://peak-performance.readthedocs.io/en/latest/) which will be placed more prominently on the new landing page. It contains a lot of helpful information which should simplify your review.
