### 2024-10-04 15e2db2
Fix typos and simplify code example
