### 2024-10-05 efb527b
Update landing page (#22)
* change landing page

* add raw data preparation to documentation

* update readme

* Fix typos and simplify code example

---------

Co-authored-by: Michael Osthege <m.osthege@fz-juelich.de>
