### 2024-10-08 #7313 @lazear
@Y0dler couple comments/suggestions/etc (will update as I go - likely to be in multiple chunks):

- You guys should be commended for the quality of your code and documentation. It's very good!
- I think the bibliography/references are messed up, only one reference appears at the end of the paper proof.
- I am not super familiar with bayesian peak picking, but I know there are other papers in this area - perhaps consider citing them and very briefly discuss differences/tradeoffs?

- Is it possible to benchmark your approach vs standard normal models on benchmark data (e.g. synthetic data with known parameters + noise?)
- Is this algorithm only suitable for analysis of a single transition in targeted LC/MS data? Could it handle multiple transitions/product ions from the same precursor?
- What are the performance characteristics of your approach? Is this scalable to integrating 1000s or 100,000s of peaks in a feasible time scale?
