### 2024-10-08 #7313 @lazear
@csoneson clarifications requested:
- Data sharing: The paper states that there is experimental data used for generating some example figures/analyses, along with a typical data availability statement: "available on reasonable request" (sic). It's unclear if this data is contained in the repository as well, since there are several pieces of example data.  I think in this case (example figure) I am fine with it, but want to check if this is in line with JOSS policies, or does the data need to be available without request?
- Contribution and authorship: the submitting author has no commits to the repository, but the first author is active in this thread. Is this OK? (I have no issues with this, submitting author appears to be the PI)
