### 2024-10-08 #7313 @Y0dler
Thanks for your comments and kind words, @lazear!

Regarding the references, I think they should be fine as far as I can see. But if you use the "View article proof on GitHub" link from editorialbot, it hides the last page with most of the references at first (until you hit the "More Pages" button at the very bottom). Maybe that was the problem? Otherwise you can also download the current draft from [GitHub actions](https://github.com/JuBiotech/peak-performance/actions/runs/11057232213).

`I am not super familiar with bayesian peak picking, but I know there are other papers in this area - perhaps consider citing them and very briefly discuss differences/tradeoffs?`
Definitely a very good point, I think we need to add that and I'll see what I can do given that we're already way past the theoretical limit of 1000 words^^ I'll try to add it as soon as possible but it might take a couple of days.

`Is it possible to benchmark your approach vs standard normal models on benchmark data (e.g. synthetic data with known parameters + noise?)`
We have tested our approach with noisy synthetic data sets of normal-shaped and skew normal-shaped peaks and compared the results with the ground truth. Originally, this was also part of the paper but we moved it to the documentation in order to decrease the word count. You can find it [here](https://peak-performance.readthedocs.io/en/latest/markdown/PeakPerformance_validation.html). I hope, this is what you meant. There is also a comparison with a commercially available vendor software using an experimental data set. To your later point regarding data sharing: this data set is currently not in the repository.

`Is this algorithm only suitable for analysis of a single transition in targeted LC/MS data? Could it handle multiple transitions/product ions from the same precursor?`
Currently, one would have to analyze multiple transitions sequentially so in case of multiple product ions, no matter if they are from the same precursor ion or not, the time series (or extracted ion chromatogram) would have to be supplied separately for each mass trace and one would have to enter the mass traces into the template Excel file so that they are analyzed in one batch run (but one after the other). I think we discussed a possibility to fit multiple peaks in a hierarchical model of sorts but for a more exact answer I would refer this one to @michaelosthege.

`What are the performance characteristics of your approach? Is this scalable to integrating 1000s or 100,000s of peaks in a feasible time scale?`
The performance depends on a number of parameters, e.g. which peak model is used for fitting, which sampler is used (ideally nutpie), whether a signal needs to be sampled again with more tuning samples or whether it is rejected right away etc.
The model selection takes a few minutes per mass trace but you only have to do that once for a batch run.
The actual peak analysis may take 20 - 30 s for a single peak and maybe 60 - 90 s for a double peak but in the notebooks it appears that most of that time is taken up by model instantiation and the sampling itself only lasts a few seconds. So, in order to scale this up to 1000s of peaks or more, one would need to parallelize the procedure on a computation cluster and/or implement a way to fit multiple peaks without re-instantiating the model every time. The latter is something that I think can be added to the software, the former is probably more on the user side. I suppose there are also ways to parallelize in Python directly but at our institute we chose the route with a computation cluster so this is not something we worked on.

Let me know if the answers were satisfactory and of course if you have more comments or questions.
