### 2024-10-11 #7313 @csoneson
Hi @lazear - thanks for your questions

* Data sharing: The paper states that there is experimental data used for generating some example figures/analyses, along with a typical data availability statement: "available on reasonable request" (sic).

According to the JOSS [data sharing policy](https://joss.readthedocs.io/en/latest/policies.html#data-sharing-policy), "If any original data analysis results are provided within the submitted work, such results have to be entirely reproducible by reviewers, including accessing the data". To me, this does seem to apply here - the figures displayed in the paper (and those comparing to the commercially available software in the documentation) were generated for the purpose of this paper.

* Contribution and authorship: the submitting author has no commits to the repository, but the first author is active in this thread. Is this OK?

We typically leave it to the authors to decide who is eligible for authorship, in line with our [authorship policy](https://joss.readthedocs.io/en/latest/submitting.html#authorship). Specifically, "Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are". Perhaps @Y0dler can comment on this for clarification.
