### 2024-10-13 2ce57dc
add paper folder with paper.md and images
