+++ paper/paper.md
@@ -1,0 +1,367 @@
+---
+title: 'PeakPerformance - a tool for Bayesian inference-based fitting of LC-MS/MS peaks'
+tags:
+  - Peak fitting
+  - Bayesian inference
+  - Chromatography
+  - LC-MS/MS
+  - Python
+  - Uncertainty quantification
+
+authors:
+  - name: Jochen Nießer
+    orcid: 0000-0001-5397-0682
+    affiliation: 1
+  - name: Michael Osthege
+    orcid: 0000-0002-2734-7624
+    affiliation: 1
+  - name: Eric von Lieres
+    orcid: 0000-0002-0309-8408
+    affiliation: "1, 2"
+  - name: Wolfgang Wiechert
+    orcid: 0000-0001-8501-0694
+    affiliation: "1, 2"
+  - name: Stephan Noack
+    orcid: 0000-0001-9784-3626
+    affiliation: 1
+
+affiliations:
+ - name: Institute for Bio- and Geosciences (IBG-1), Forschungszentrum Jülich GmbH, Jülich, Germany
+   index: 1
+ - name: Computational Systems Biotechnology, RWTH Aachen University, Aachen, Germany
+   index: 2
+
+date: 9th of August 2024
+bibliography: literature.bib
+
+# Summary
+
+A major bottleneck of chromatography-based analytics has been the elusive fully automated identification and integration of peak data without the need of extensive human supervision.
+The presented Python package $\texttt{PeakPerformance}$ applies Bayesian inference to chromatographic peak fitting, and provides an automated approach featuring model selection and uncertainty quantification.
+Currently, its application is focused on data from targeted liquid chromatography tandem mass spectrometry (LC-MS/MS), but its design allows for an expansion to other chromatographic techniques.
+$\texttt{PeakPerformance}$ is implemented in Python and the source code is available on [GitHub](https://github.com/JuBiotech/peak-performance).
+It is unit-tested on Linux and Windows and accompanied by general introductory documentation, as well as example notebooks.
+
+# Statement of need
+
+In biotechnological research and industrial applications, chromatographic techniques are ubiquitously used to analyze samples from fermentations, e.g. to determine the concentration of substrates and products.
+Over the course of a regular lab-scale bioreactor fermentation, dozens to hundreds of samples and subsequently – depending on the number of analytes per sample – hundreds to thousands of chromatographic peaks may accrue.
+This is exacerbated by the spread of microbioreactors causing a further increase in the amount of samples per time [@RN149, @RN148].
+While the recognition and integration of peaks by vendor software is – in theory – automated, it typically requires visual inspection and occasional manual re-integration by the user due to a large number of false positives, false negatives or incorrectly determined baselines, ultimately downgrading it to a semi-automated process.
+Since this is a time-consuming, not to mention tedious, procedure and introduces the problem of comparability between purely manual and algorithm-based integration as well as user-specific differences, we instead propose a peak fitting solution based on Bayesian inference.  
+The advantage of this approach is the complete integration of all relevant parameters – i.e. baseline, peak area and height, mean, signal-to-noise ratio etc. – into one single model through which all parameters are estimated simultaneously.
+Furthermore, Bayesian inference comes with uncertainty quantification for all peak model parameters, and thus does not merely yield a point estimate as would commonly be the case.
+It also grants access to novel metrics for avoiding false positives and negatives by rejecting signals where a) a convergence criterion of the peak fitting procedure was not fulfilled or b) the uncertainty of the estimated parameters exceeded a user-defined threshold.
+
+# Materials and Methods
+## Implementation
+$\texttt{PeakPerformance}$ is an open source Python package compatible with Windows and Linux/Unix platforms.
[…]
