### 2024-10-13 8416645
test removing begin and end equation
