+++ paper/paper.md
@@ -102,41 +102,35 @@ If specific distributions or their parameters had to be restricted to certain va
 For example, when only positive values were acceptable or when 0 was not a permissive value, a lower bound was defined using NumPy's clip function.  
 
 Regarding shared model elements across all intensity functions, one such component of all models presented hereafter is the likelihood function
-$$\begin{equation}\tag{1}
-    L \sim Normal(y, noise)
-\end{equation}$$
+$$\tag{1}L \sim Normal(y, noise)$$
 with $y$ as the predicted intensity and $noise$ as the free parameter describing the standard deviation of measurement noise.
 This definition encodes the assumption that observed intensities are the result of normally distributed noise around the true intensity values of a peak.
 In turn, the noise parameter is defined as 
-$$\begin{equation}\tag{2}
-    noise \sim LogNormal(\log_{10} max(10, noise_{guess}), 1)
-\end{equation}$$
+$$\tag{2}noise \sim LogNormal(\log_{10} max(10, noise_{guess}), 1)$$
 The log-normal distribution where the logarithm of the random variable follows a normal distribution was chosen partly to exclude negative values from the solution space and also due to its shape attributing a higher fraction of the probability density to lower values provided the standard deviation is defined sufficiently high.
 This prior is defined in a raw data-dependent manner as the $noise_{guess}$ amounts to the standard deviation of the differences of the first and final 15 \% of intensity values included in a given time frame and their respective mean values.  
 
 The intensity function itself is defined as the sum of a linear baseline function and a peak intensity function, the latter of which is composed of a given distribution's probability density function (PDF) scaled up to the peak size by the area or height parameter.
 The linear baseline
-$$\begin{equation}\tag{3}
-    y_{baseline}(t) = at+b
-\end{equation}$$
+$$\tag{3}y_{baseline}(t) = at+b$$
 features the slope and intersect parameters $a$ and $b$, respectively, both of which were given a normally distributed prior.
 The data-dependent guesses for these priors are obtained by constructing a line through the means of the first and last three data points of a given intensity data set which oftentimes already resulted in a good fit.
 Hence, the determined values for slope ($a_{guess}$) and intercept ($b_{guess}$) are used as the means for their pertaining priors and the standard deviations are defined as fractions of them with minima set to 0.5 and 0.05, respectively.
 Here, the exact definition of the standard deviations was less important than simply obtaining an uninformative prior which, while based on the rough fit for the baseline, possesses a sufficient degree of independence from it, thus allowing deviations by the Bayesian parameter estimation.
-$$\begin{equation}\tag{4}
+$$\tag{4}
     a \sim 
     \begin{cases}
         Normal(a_{guess}, \frac{|a_{guess}|}{5})   &  if\ \frac{|a_{guess}|}{5}\geq0.5\\
         Normal(a_{guess}, 0.5)                   & otherwise\\
     \end{cases}
-\end{equation}$$
-$$\begin{equation}\tag{5}
+$$
+$$\tag{5}
     b \sim 
     \begin{cases}
         Normal(b_{guess}, \frac{|b_{guess}|}{6})   & if\ \frac{|b_{guess}|}{6}\geq0.05\\
         Normal(b_{guess}, 0.05)                   & otherwise\\
     \end{cases}
-\end{equation}$$
+$$
[…]
