### 2024-10-13 9d7ec19
change $ to $$ in table 1
