### 2024-10-13 d24bf42
remove stale reference to equation 6
