+++ paper/paper.md
@@ -287,16 +287,15 @@ To complete the example, Table 2 shows the results of the fit in the form of mea
 __Table 2:__ Depiction of the results for the most important peak parameters of a single peak fit with the skew normal model and a double peak fit with the double normal model. Mean, area, and height have been highlighted in bold print as they constitute the most relevant parameters for further data evaluation purposes. The results correspond to the fits exhibited in Figure 4.
 ![](./summary_joint.png)
 
-+---------------------+-----------------------+
-| Location            | Temperature 1961-1990 |
-|                     | in degree Celsius     |
-|                     +-------+-------+-------+
-|                     | min   | mean  | max   |
-+=====================+=======+=======+=======+
-| Antarctica          | -89.2 | N/A   | 19.8  |
-+---------------------+-------+-------+-------+
-| Earth               | -89.2 | 14    | 56.7  |
-+---------------------+-------+-------+-------+
++---------------------+----------+
+| Property            | Earth    |
++=============+=======+==========+
+|             | min   | -89.2 °C |
+| Temperature +-------+----------+
+| 1961-1990   | mean  | 14 °C    |
+|             +-------+----------+
+|             | max   | 56.7 °C  |
++-------------+-------+----------+
 
 In this case, the fits were successful and convergence was reached for all parameters.
 Most notably and for the first time, the measurement noise was taken into account when determining the peak area as represented by its standard deviation and as can be observed in the posterior predictive plots where the noisy data points fall within the boundary of the 95 \% HDI.
