### 2024-10-13 c2cb84b
move Figure 2 and final additions of mathrm
