### 2024-10-13 3f3dbc2
remove dot behind equation
