### 2024-10-13 9263f60
Format table 2 as pandoc grid table
If this does not work, rever to
```
![](./summary_joint.svg){width="100%"}
```
