+++ paper/paper.md
@@ -287,7 +287,35 @@ Accordingly, this plot enables users to judge whether the selected model can acc
 To complete the example, Table 2 shows the results of the fit in the form of mean, standard deviation, and HDI of each parameter's marginal posterior.
 
 __Table 2:__ Depiction of the results for the most important peak parameters of a single peak fit with the skew normal model and a double peak fit with the double normal model. Mean, area, and height have been highlighted in bold print as they constitute the most relevant parameters for further data evaluation purposes. The results correspond to the fits exhibited in Figure 4.
-![](./summary_joint.svg)
+
++--------------------+-----------------------------------------------------------------+-------------------------------------------------------------------+
+| parameter          | single peak (skew normal model)                                 | double peak (double normal model)                                 |
++--------------------+---------------------------------+-----------+---------+---------+-----------------------------------+-----------+---------+---------+
+|                    | mean                            | sd        | hdi_3%  | hdi_97% | mean                              | sd        | hdi_3%  | hdi_97% |
++:===================+================================:+==========:+========:+========:+==================================:+==========:+========:+========:+
+| baseline_intercept |                          -43.94 |      7.41 |  -57.88 |  -30.02 |                           1115.40 |     38.69 | 1040.14 | 1185.07 |
++--------------------+---------------------------------+-----------+---------+---------+-----------------------------------+-----------+---------+---------+
+| baseline_slope     |                            6.66 |      0.51 |    5.71 |    7.63 |                            -21.65 |      3.09 |  -27.50 |  -15.94 |
++--------------------+---------------------------------+-----------+---------+---------+-----------------------------------+-----------+---------+---------+
+| noise              |                          103.63 |      7.51 |    89.5 |  117.26 |                            118.63 |      8.01 |  103.52 |  133.29 |
++--------------------+---------------------------------+-----------+---------+---------+-----------------------------------+-----------+---------+---------+
+| mean               |                       __25.95__ |  __0.01__ |   25.93 |   25.97 |                         __11.73__ |  __0.02__ |   11.70 |   11.76 |
+|                    |                                 |           |         |         |                         __12.43__ |  __0.01__ |   12.42 |   12.45 |
++--------------------+---------------------------------+-----------+---------+---------+-----------------------------------+-----------+---------+---------+
+| area               |                     __1512.32__ | __37.31__ | 1141.25 | 1581.37 |                        __317.16__ | __28.84__ | 263.23  |  370.56 |
+|                    |                                 |           |         |         |                        __674.34__ | __26.34__ | 623.47  |  722.88 |
++--------------------+---------------------------------+-----------+---------+---------+-----------------------------------+-----------+---------+---------+
+| height             |                     __1879.72__ | __37.71__ | 1809.30 | 1950.64 |                        __774.99__ | __65.28__ | 653.50  |  897.88 |
+|                    |                                 |           |         |         |                       __1762.66__ | __64.04__ | 1639.62 | 1881.66 |
++--------------------+---------------------------------+-----------+---------+---------+-----------------------------------+-----------+---------+---------+
+| std                |                            0.53 |      0.02 |    0.48 |    0.56 |                              0.16 |      0.02 |    0.13 |    0.20 |
+|                    |                                 |           |         |         |                              0.15 |      0.01 |    0.14 |    0.17 |
++--------------------+---------------------------------+-----------+---------+---------+-----------------------------------+-----------+---------+---------+
+| sn                 |                           18.24 |      1.37 |   15.69 |   20.76 |                              6.56 |      0.72 |    5.22 |    7.88 |
+|                    |                                 |           |         |         |                             14.93 |      1.14 |   12.82 |   17.14 |
[…]
