### 2024-10-13 2f28e9e
fix percent signs, some citations, and R_hat
