### 2024-10-13 ef1e5e0
Revert to embedding Table 2 from SVG
