+++ paper/literature.bib
@@ -192,12 +192,12 @@
 
 @article{harris2020array,
 	title         = {Array programming with {NumPy}},
-	author        = {Harris, C. R. and Millman, K. J. and 
+	author        = {Harris, C. R. and Millman, K. J. and
 	{van der Walt}, S. J. and Gommers, R. and Virtanen, P. and
-	Cournapeau, D. and Wieser, E. and Taylor, J. and 
+	Cournapeau, D. and Wieser, E. and Taylor, J. and
 	Berg, S. and Smith, N. J. and Kern, R. and Picus, M.
-	and Hoyer, S. and {van Kerkwijk}, M. H. and 
-	Brett, M. and Haldane, M. and del R{\'{i}}o, J. F. and Wiebe, M. and Peterson, P. and 
+	and Hoyer, S. and {van Kerkwijk}, M. H. and
+	Brett, M. and Haldane, M. and del R{\'{i}}o, J. F. and Wiebe, M. and Peterson, P. and
 	G{\'{e}}rard-Marchant, P. and Sheppard, K. and Reddy, T. and
 	Weckesser, W. and Abbasi, H. and Gohlke, C. and
 	Oliphant, T. E.},
