### 2024-10-13 0ab5e62
fix matplotlib reference and add numpy one
