+++ paper/literature.bib
@@ -45,7 +45,7 @@
 }
 
 @softmisc{matplotlibzenodo,
-  author    = {The Matplotlib Development Team},
+  author    = {{The Matplotlib Development Team}},
   title     = {Matplotlib: Visualization with Python},
   keywords  = {software},
   month     = may,
@@ -188,4 +188,26 @@
   doi     = {10.21105/joss.01143},
   year    = {2019},
   type    = {Journal Article}
+}
+
+@article{harris2020array,
+	title         = {Array programming with {NumPy}},
+	author        = {Harris, C. R. and Millman, K. J. and 
+	{van der Walt}, S. J. and Gommers, R. and Virtanen, P. and
+	Cournapeau, D. and Wieser, E. and Taylor, J. and 
+	Berg, S. and Smith, N. J. and Kern, R. and Picus, M.
+	and Hoyer, S. and {van Kerkwijk}, M. H. and 
+	Brett, M. and Haldane, M. and del R{\'{i}}o, J. F. and Wiebe, M. and Peterson, P. and 
+	G{\'{e}}rard-Marchant, P. and Sheppard, K. and Reddy, T. and
+	Weckesser, W. and Abbasi, H. and Gohlke, C. and
+	Oliphant, T. E.},
+	year          = {2020},
+	month         = sep,
+	journal       = {Nature},
+	volume        = {585},
+	number        = {7825},
+	pages         = {357--362},
+	doi           = {10.1038/s41586-020-2649-2},
+	publisher     = {Springer Science and Business Media {LLC}},
+	url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
