### 2024-10-13 5eccece
shorten even more, remove diagnostic plots
