+++ paper/paper.md
@@ -44,12 +44,12 @@ A major bottleneck of chromatography-based analytics has been the elusive fully 
 The presented Python package $\texttt{PeakPerformance}$ applies Bayesian inference to chromatographic peak fitting, and provides an automated approach featuring model selection and uncertainty quantification.
 Currently, its application is focused on data from targeted liquid chromatography tandem mass spectrometry (LC-MS/MS), but its design allows for an expansion to other chromatographic techniques.
 $\texttt{PeakPerformance}$ is implemented in Python and the source code is available on [GitHub](https://github.com/JuBiotech/peak-performance).
-It is unit-tested on Linux and Windows and accompanied by general introductory documentation, as well as example notebooks.
+It is unit-tested on Linux and Windows and accompanied by documentation as well as example notebooks.
 
 # Statement of need
 
 In biotechnological research and industrial applications, chromatographic techniques are ubiquitously used to analyze samples from fermentations, e.g. to determine the concentration of substrates and products.
-Over the course of a regular lab-scale bioreactor fermentation, dozens to hundreds of samples and subsequently – depending on the number of analytes per sample – hundreds to thousands of chromatographic peaks may accrue.
+Over the course of a regular lab-scale bioreactor fermentation, hundreds of samples and subsequently thousands of chromatographic peaks may accrue.
 This is exacerbated by the spread of microbioreactors causing a further increase in the amount of samples per time [@RN149; @RN148].
 While the recognition and integration of peaks by vendor software is – in theory – automated, it typically requires visual inspection and occasional manual re-integration by the user due to a large number of false positives, false negatives or incorrectly determined baselines, ultimately downgrading it to a semi-automated process.
 Since this is a time-consuming, not to mention tedious, procedure and introduces the problem of comparability between purely manual and algorithm-based integration as well as user-specific differences, we instead propose a peak fitting solution based on Bayesian inference.
@@ -63,8 +63,7 @@ $\texttt{PeakPerformance}$ is an open source Python package compatible with Wind
 At the time of manuscript submission, it features three modules: `pipeline`, `models`, and `plotting`.
 Due to its modular design, $\texttt{PeakPerformance}$ can easily be expanded by adding e.g. additional models for deviating peak shapes or different plots.
 Currently, the featured peak models describe peaks in the shape of normal or skew normal distributions, as well as double peaks of normal or skewed normal shape.
-The normal distribution is regarded as the ideal peak shape and common phenomena like tailing and fronting can be expressed by the skew normal distribution [@RN144].
-A third probability density function commonly used to describe peak shapes is the PDF of the exponentially modified Gaussian [@RN153] for which an implementation in $\texttt{PeakPerformance}$ is pending.\\
+The normal distribution is regarded as the ideal peak shape and common phenomena like tailing and fronting can be expressed by the skew normal distribution [@RN144].\\
 Bayesian inference is conducted utilizing the PyMC package [@RN150] with the external sampler $\texttt{nutpie}$ for improved performance [@nutpie].
 Both model selection and analysis of inference data objects are realized with the ArviZ package [@RN147].
 Since the inference data is stored alongside graphs and report sheets, users may employ the ArviZ package or others for further analysis of the results if necessary.
@@ -94,56 +93,31 @@ The data set consisted of 192 signals comprised of 123 single peaks, 50 peaks as
 
 ## PeakPerformance workflow
 $\texttt{PeakPerformance}$ accommodates the use of a pre-manufactured data pipeline for standard applications (Fig. 1) as well as the creation of custom data pipelines using only its core functions.
-The provided data analysis pipeline was designed in a user-friendly way and requires minimal programming knowledge.
-As portrayed in an example notebook in the code repository, only a few simple Python commands need to be executed.
[…]
