### 2024-10-13 4339aed
Restructure documentation index
Switched from RestructuredText to MyST because it's easier.
Created sections in the left-hand sidebar as discussed.
