### 2024-10-13 fc2d48f
Fix formatting of license and module docstrings
