### 2024-10-13 ce3eafc
Fix various documentation glitches / warnings
