+++ peak_performance/models.py
@@ -29,12 +29,19 @@ import scipy.stats as st
 
 
 class ModelType(str, Enum):
-    """Class containing all implemented model types."""
+    """Enum of default model types."""
 
     Normal = "normal"
+    """Shape of a Gaussian Normal PDF."""
+
     SkewNormal = "skew_normal"
+    """Shape of a skewed Normal PDF."""
+
     DoubleNormal = "double_normal"
+    """Superposition of two ``Normal`` peaks."""
+
     DoubleSkewNormal = "double_skew_normal"
+    """Superposition of two ``SkewedNormal`` peaks."""
 
 
 def guess_noise(intensity):
@@ -393,10 +400,9 @@ def std_skew_calculation(scale, alpha):
         Skewness parameter of the skew normal distribution.
 
     Returns
-    ----------
+    -------
     std
         Standard deviation of a skew normal distribution.
-    -------
     """
     return np.sqrt(scale**2 * (1 - (2 * alpha**2) / ((alpha**2 + 1) * np.pi)))
 
@@ -415,7 +421,7 @@ def mean_skew_calculation(loc, scale, alpha):
         Skewness parameter of the skew normal distribution.
 
     Returns
-    ----------
+    -------
     mean
         Arithmetic mean of a skew normal distribution.
     """
@@ -489,7 +495,7 @@ def height_calculation(area, loc, scale, alpha, mode_skew):
         Mode of the skew normal distribution.
 
     Returns
-    ----------
+    -------
     mean
         Arithmetic mean of a skew normal distribution.
     """
