### 2024-10-13 395b89c
add environment.yml
