### 2024-10-13 5617d6c
Update docs/source/markdown/Installation.md
Co-authored-by: Michael Osthege <m.osthege@fz-juelich.de>
