### 2024-10-13 f14af7a
bump version number and change data availability statement
