### 2024-10-13 #7313 @Y0dler
Hi @csoneson,

regarding data sharing: the repo was updated so that the notebooks to reproduce the data processing etc. have been added to `peak-performance\docs\source\notebooks` and the raw data to `peak-performance\docs\source\notebooks\paper raw data` and its sub-directories. In case the repo changes down the road, the raw data can be found on [zenodo](https://zenodo.org/records/13925914) under version 0.7.1. The data availability statement was changed accordingly.

Regarding the authorship, @MicroPhen did indeed contribute via active project direction. For example, it was his idea to perform the validation via synthetic data, he helped identify a bug with the posterior predictive sampling etc. Also, I have since moved on to a different job, so further developments of the software will be realized by a successor under the guidance of @MicroPhen.

Also, we merged the paper branch into the main branch and created a version 0.7.1 release. This does not contain any changes to the program code, it just adds the paper, raw data and so on to the repo. Also, we changed the instructions for the installation (@lazear, @Adafede) since there were version conflicts with NumPy, PyMC, and numba.
