### 2024-10-13 #7313 @michaelosthege
>> Is this algorithm only suitable for analysis of a single transition in targeted LC/MS data? Could it handle multiple transitions/product ions from the same precursor?

> I think we discussed a possibility to fit multiple peaks in a hierarchical model of sorts but for a more exact answer I would refer this one to @michaelosthege.

Yes, using the functions from `peak_performance.models` it is possible to define a hierarchical PyMC model, that could connect multiple peaks by, for example, sharing the retention time parameter.
This would enable a more wholistic quantification of uncertainties and is something we would like to explore in the future, but it was out of scope for the current release of the library.
