### 2024-10-17 #7313 @lazear
I've tested installation and run the examples from the documentation. Besides a [minor versioning issue](https://github.com/JuBiotech/peak-performance/issues/27) everything went smoothly. I think all of the checklist items for me are complete, and again, the authors should be commended on the excellent state of the documentation, code quality, and ease of installation and verification. I recommend publication.

My only remaining comments are that it would be nice to briefly mention performance/runtime of the algorithm in the paper as above.
