### 2024-10-17 JuBiotech/peak-performance#27 @michaelosthege
This is a bug I've seen multiple times this week, with different projects. I suspect the NumPy pins by PyTensor are not strict enough for all install methods.
I couldn't reproduce it this time, but will nevertheless pin `numpy <2` and remove it once PyTensor becomes compatible.

Thanks for pointing it out!
