### 2024-10-17 df5bbf6
Pin NumPy 1 for the time being
Closes #27
