### 2024-10-19 #7313 @Y0dler
Thanks for the review! I'll add your suggested additions regarding the performance of the software and the comparison to other papers applying Bayesian statistic to peak data evaluation to the pull request you opened 👍
