### 2024-10-21 JuBiotech/peak-performance#31 @Adafede
Improve first user experience
Dear @Y0dler, and @michaelosthege

I have been assigned to review your repository for https://github.com/openjournals/joss-reviews/issues/7313.

As this is my first issue, please be ensured I highly appreciate **a lot** of points that are already available. Per convenience, I will not list all the things that are already done well, but only the "issues" so this might give a way more negative impression than what my real thoughts are, my apoogies for it already!

- [x] Add link to the automatically generated documentation as repository link in the "about" section
- [x] The README does not contain much and directly redirects to the doc. The very basic initial steps should be in there, more or less like the first page of the doc.
- [x] Admitting a user does not want to install the software using pip, are there other alternatives? Could alternative installations (poetry, mamba, ...) be made available for users? Oh... this was answered on the next page, probably some other lazy users will also not want to get to page 2 to read if they can install it without using pip?
- [x] Provide a full example for https://peak-performance.readthedocs.io/en/latest/markdown/Preparing_raw_data.html#preparing-raw-data and not let users have to create lists for time and intensity on their own?
