### 2024-10-21 JuBiotech/peak-performance#32 @Adafede
Handle outputs from classical peak integration pipelines
In the frame of my review for https://github.com/openjournals/joss-reviews/issues/7313.

This is more of a discussion rathen than a hard requirement...

I work with LC-MS data processing softwares on a daily basis, and to really get your repository out of its actual "niche", I believe that handling directly _some_ of the classical outputs from other softwares would help a lot!

While the ".xlsx" way to do things might help some users, if others really want to integrate your pipeline into larger ones, they will need some convenience scripts to generate these from their usually used softwares, do you think handling some examples might be doable?
