### 2024-10-21 JuBiotech/peak-performance#33 @Adafede
Improve notebooks organization
In the frame of my review for https://github.com/openjournals/joss-reviews/issues/7313.

When I landed on https://github.com/JuBiotech/peak-performance/tree/main/docs/source/notebooks, of course there is everything I need to reproduce but having some clearer instinctive way to know where to start might help (like `01_create_x`, `02_do_this`, `03_final_that` ...)

On a more personal note, https://marimo.io/ recently got me much more enthusiastic than jupyter notebooks
