### 2024-10-21 JuBiotech/peak-performance#34 @Adafede
Unusual `CONTRIBUTING.md`?
In the frame of my review for https://github.com/openjournals/joss-reviews/issues/7313.

I am used to repositories using the classical `CONTRIBUTING.md` adopted by GitHub, and not `CLA`. Is there any particular reason for using it? I genuinely don't know and so maybe this is fully justified, just felt unusual.
