### 2024-10-21 #7313 @Adafede
Some additional notes:
- [x] **Contribution and authorship**: This has been clarified above
- [x] **Data sharing**: Fixed thanks to @lazear
- [ ] **Repository**: See https://github.com/JuBiotech/peak-performance/issues/31
- [ ] **Reproducibility**: See https://github.com/JuBiotech/peak-performance/issues/33
- [ ] **Community guidelines**: See https://github.com/JuBiotech/peak-performance/issues/34
- [ ] **State of the field**: It is unclear to me how to exactly "fit" the author's pipeline into actual pipelines (see related https://github.com/JuBiotech/peak-performance/issues/32)
