### 2024-10-24 JuBiotech/peak-performance#33 @Y0dler
I added such a new entry in #35.
