### 2024-10-24 JuBiotech/peak-performance#31 @Y0dler
By basic steps, do you refer just to the installation instructions?

I updated the about section with the link to the documentation and changed the landing page of the docu to contain all installation instructions in a new PR.
