### 2024-10-24 7714a4c
add installation instructions to README.md
