### 2024-10-24 JuBiotech/peak-performance#31 @Adafede
Yes, you adressed everything except the example time and intensity arrays in #35  😊
