### 2024-10-24 3e78028
move complete installation instructions to docu's landing page
