### 2024-10-24 a058ed0
add docu section about recreating results from raw data
