### 2024-10-25 be84ab8
Create notebook for preparing raw data with more extensive explanation
