### 2024-10-25 JuBiotech/peak-performance#31 @Y0dler
Just replaced the markdown file with a Jupyter notebook which shows how to open a `*.npy` file (although I guess this was also covered in some of the other notebooks) and then gives an example of an actual metabolite and a random time series showing how the file would have to be named and how it can be stored. Is this sufficient?
