### 2024-10-25 28f3e7c
Update paper.md with performance
