### 2024-10-25 5bcb6ea
Update literature.bib
