+++ docs/source/literature.bib
@@ -212,3 +212,78 @@
 	publisher     = {Springer Science and Business Media {LLC}},
 	url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
+
+@article{vivo2012bayesian,
+  title={Bayesian approach for peak detection in two-dimensional chromatography},
+  author={Viv{\'o}-Truyols, Gabriel},
+  journal={Analytical chemistry},
+  volume={84},
+  number={6},
+  pages={2622--2630},
+  year={2012},
+  publisher={ACS Publications}
+}
+
+@article{woldegebriel2015probabilistic,
+  title={Probabilistic model for untargeted peak detection in LC--MS using Bayesian statistics},
+  author={Woldegebriel, Michael and Viv{\'o}-Truyols, Gabriel},
+  journal={Analytical chemistry},
+  volume={87},
+  number={14},
+  pages={7345--7355},
+  year={2015},
+  publisher={ACS Publications}
+}
+
+@article{briskot2019prediction,
+  title={Prediction uncertainty assessment of chromatography models using Bayesian inference},
+  author={Briskot, Till and St\"{u}ckler, Ferdinand and Wittkopp, Felix and Williams, Christopher and Yang, Jessica and Konrad, Susanne and Doninger, Katharina and Griesbach, Jan and Bennecke, Moritz and Hepbildikler, Stefan and others},
+  journal={Journal of Chromatography A},
+  volume={1587},
+  pages={101--110},
+  year={2019},
+  publisher={Elsevier}
+}
+
+@article{yamamoto2021uncertainty,
+  title={Uncertainty quantification for chromatography model parameters by Bayesian inference using sequential Monte Carlo method},
+  author={Yamamoto, Yota and Yajima, Tomoyuki and Kawajiri, Yoshiaki},
+  journal={Chemical Engineering Research and Design},
+  volume={175},
+  pages={223--237},
+  year={2021},
+  publisher={Elsevier}
+}
+
+@article{wiczling2016much,
+  title={How much can we learn from a single chromatographic experiment? A Bayesian perspective},
+  author={Wiczling, Pawe{\l} and Kaliszan, Roman},
+  journal={Analytical chemistry},
+  volume={88},
+  number={1},
+  pages={997--1002},
+  year={2016},
+  publisher={ACS Publications}
+}
+
+@article{kelly1971estimation,
+  title={Estimation of chromatographic peaks with particular consideration of effects of base-line noise},
+  author={Kelly, PC and Harris, WE},
+  journal={Analytical Chemistry},
+  volume={43},
+  number={10},
+  pages={1170--1183},
+  year={1971},
+  publisher={ACS Publications}
+}
+
+@article{kelly1971application,
+  title={Application of method of maximum posterior probability to estimation of gas-chromatographic peak parmeters},
+  author={Kelly, PC and Harris, WE},
+  journal={Analytical Chemistry},
+  volume={43},
+  number={10},
+  pages={1184--1195},
+  year={1971},
+  publisher={ACS Publications}
+}
