### 2024-10-25 e2b5378
Update paper.md with short comparison to other publications
