### 2024-10-25 #7313 @Y0dler
@lazear - I just added a sentence addressing the performance of the software as well as a very short comparison to previous literature to your PR. In case you are fine with these changes and sign the CLA, we can go ahead and merge it.

@Adafede - I also introduced changes to [our PR](https://github.com/JuBiotech/peak-performance/pull/35) with regards to the preparation of `*.npy` for the program. Instead of the markdown document, it is now a notebook with a little more detailed explanations.
