### 2024-10-25 JuBiotech/peak-performance#31 @Adafede
@Y0dler Perfect, it will probably help a lot of beginners 😉
