### 2024-10-25 #7313 @Adafede
@Y0dler Thank you for all these!

I admit https://github.com/JuBiotech/peak-performance/issues/32 is probably quite hard so maybe @lazear or @csoneson can comment if this seems reasonable?
