### 2024-10-26 8619097
Minor syntax fixes
