### 2024-10-26 569c146
Improve Documentation UX (#35)
* move complete installation instructions to docu's landing page
* results plots are now shown in notebook instead of being stored
* add docu section about recreating results from raw data
* apply pre-commit changes
* add installation instructions to README.md
* Create notebook for preparing raw data with more extensive explanation
* change index accordingly

Closes #31

---------

Co-authored-by: Michael Osthege <m.osthege@fz-juelich.de>
