+++ README.md
@@ -9,6 +9,28 @@ PeakPerformance employs Bayesian modeling for chromatographic peak data fitting.
 This has the innate advantage of providing uncertainty quantification while jointly estimating all peak parameters united in a single peak model.
 As Markov Chain Monte Carlo (MCMC) methods are utilized to infer the posterior probability distribution, convergence checks and the aformentioned uncertainty quantification are applied as novel quality metrics for a robust peak recognition.
 
+# Installation
+
+It is highly recommended to follow the following steps and install ``PeakPerformance`` in a fresh Python environment:
+1. Install the package manager [Mamba](https://github.com/conda-forge/miniforge/releases).
+Choose the latest installer at the top of the page, click on "show all assets", and download an installer denominated by "Mambaforge-version number-name of your OS.exe", so e.g. "Mambaforge-23.3.1-1-Windows-x86_64.exe" for a Windows 64 bit operating system. Then, execute the installer to install mamba and activate the option "Add Mambaforge to my PATH environment variable".
+
+⚠ If you have already installed Miniconda, you can install Mamba on top of it but there are compatibility issues with Anaconda.
+
+ℹ The newest conda version should also work, just replace `mamba` with `conda` in step 2.
+
+2. Create a new Python environment in the command line using the provided [`environment.yml`](https://github.com/JuBiotech/peak-performance/blob/main/environment.yml) file from the repo.
+   Download `environment.yml` first, then navigate to its location on the command line interface and run the following command:
+```
+mamba env create -f environment.yml
+```
+
+Naturally, it is alternatively possible to just install ``PeakPerformance`` via pip:
+
+```bash
+pip install peak-performance
+```
+
 # First steps
 Be sure to check out our thorough [documentation](https://peak-performance.readthedocs.io/en/latest). It contains not only information on how to install PeakPerformance and prepare raw data for its application but also detailed treatises about the implemented model structures, validation with both synthetic and experimental data against a commercially available vendor software, exemplary usage of diagnostic plots and investigation of various effects.
 Furthermore, you will find example notebooks and data sets showcasing different aspects of PeakPerformance.
