### 2024-11-03 #7313 @csoneson
Regarding https://github.com/JuBiotech/peak-performance/issues/32 - I think providing an explanation and/or some examples of how to generate the required input files from 'standard' files available to the user would indeed help adoption by people who don't have access to the custom data preparation pipeline used by the authors. Perhaps in this case the `mzML` format can be considered a suitable standardized starting point. Even without an automatic conversion function in place, I would suggest adding some examples (which can then be adapted by the user if needed) that walk through the process of parsing these files and saving the data in a suitable format for further analysis.
