### 2024-11-03 #7313 @lazear
I agree that It might be nice having an example highlighting mzML usage, but IMO given that this package is a relatively niche library and not necessarily a standalone tool, it seems somewhat out of scope to me to require the authors to develop a full mzML pipeline given that this is largely for processing SRM/single ion data. They take numpy arrays as input, which can be easily obtained from mzML files (pyteomics or pymzML python packages) - presumably anyone using this library will be able to manage loading data into numpy arrays.
