### 2024-11-06 #7313 @Y0dler
@Adafede, @lazear, @csoneson - I added such an example (see [PR 36](https://github.com/JuBiotech/peak-performance/pull/36)). It starts from a `*.wiff` file (Sciex) which I converted to `*mzML` via ProteoWizard. Subsequently, I expanded the `Preparing raw data...` example to include a section where the file is opened and an extracted ion chromatogram is parsed via `pyteomics` and prepared for use with `PeakPerformance`.

However, I can't upload the mzML file to GitHub since it is huge (158 MB). We'll have to find a different solution for that.
