### 2024-11-08 #7313 @Y0dler
Okay, a solution was found and the `*.mzML` file was attached to release [0.7.1](https://github.com/JuBiotech/peak-performance/releases/tag/v0.7.1). The instructions in the example notebook were updated accordingly, as well.

I think the final open point remains [PR 28](https://github.com/JuBiotech/peak-performance/pull/28) which can only be merged once @lazear has signed the CLA. Sorry about that. If you would prefer not to sign it, I can create a new PR, instead, and credit you in it :)
