### 2024-11-08 9452057
update literature.bib
