### 2024-11-08 050d9d7
add missing entry in literature.bib
