+++ docs/source/literature.bib
@@ -255,6 +255,17 @@
   publisher={Elsevier}
 }
 
+@article{wiczling2016much,
+  title={How much can we learn from a single chromatographic experiment? A Bayesian perspective},
+  author={Wiczling, Pawe{\l} and Kaliszan, Roman},
+  journal={Analytical chemistry},
+  volume={88},
+  number={1},
+  pages={997--1002},
+  year={2016},
+  publisher={ACS Publications}
+}
+
 @article{kelly1971estimation,
   title={Estimation of chromatographic peaks with particular consideration of effects of base-line noise},
   author={Kelly, PC and Harris, WE},
