### 2024-11-11 c9f7c21
move figure 3
