### 2024-11-11 #7313 @Y0dler
Yeah, it's weird :D even after re-running the CLA job, it didn't work so I went with a new PR instead and closed the old one.

@csoneson I think you can trigger the generation of a new PDF here so we can check if everything compiles correctly (in GitHub Actions it did so I assume there won't be a problem).
