### 2024-11-11 #7313 @csoneson
@Y0dler You can also trigger the pdf generation on your side (with `@editorialbot generate pdf`)
