### 2024-11-11 #7313 @Adafede
@csoneson Is it expected that editors get their ORCID linked but not the reviewers?
