### 2024-11-13 #7313 @csoneson
> @csoneson Is it expected that editors get their ORCID linked but not the reviewers?

Yes, this is the case currently. It has been discussed in the editorial team, and linking reviewer ORCID would likely be technically possible (if the reviewers have added it in their JOSS reviewer profile), but the development resources are limited and other aspects have had priority.
