### 2024-11-19 #7313 @csoneson
Sorry for the delay. I see that both reviewers have checked all boxes in the checklists, and @lazear recommended publication in a previous comment. @Adafede - are there any outstanding comments on your side?

Once we have the explicit approval from both reviewers, I will also take a quick look through the submission, and then we can move on.
