### 2024-11-25 #7313 @csoneson
👋🏻 @Y0dler - I have gone through the submission and have a few comments related to the paper:

- First, as we have discussed before, it would be great if it could be shortened a bit more to adhere better to the JOSS guidelines (it's actually longer now than when then review started 🙂). I think much of the detailed explanations in the Results and Discussion part can be moved to a separate documentation file, as this is important when the package is used in practice, but perhaps not essential to give an idea of its capabilities. Also the Discussion part is largely already covered in the initial Summary. [Link to JOSS guidelines](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain)
- line 39: there is an extra ,
- line 46: methodolody -> methodology
- line 46: quiet -> quite
- line 69: remove 'according'
- line 143: remove \
- in the bibliography, please spell out journal names
- also please remove [Journal Article] from references
- there is some case protection needed in some of the references
- in Kelly & Harris 1971a: parmeters -> parameters
- I'm not sure why the reference check above is not responsive at the moment, but there are quite a few references without DOIs - are there any of those that can be added?

Thanks!
