### 2024-11-25 98731ef
fix typos, update bibliography
