+++ paper/literature.bib
@@ -1,4 +1,4 @@
-@softmisc{nutpie,
+@misc{nutpie,
   author   = {Seyboldt, Adrian and {PyMC Developers}},
   keywords = {Software},
   license  = {MIT},
@@ -44,7 +44,7 @@
   year      = 2007
 }
 
-@softmisc{matplotlibzenodo,
+@misc{matplotlibzenodo,
   author    = {{The Matplotlib Development Team}},
   title     = {Matplotlib: Visualization with Python},
   keywords  = {software},
@@ -62,7 +62,6 @@
   journal = {Journal of Machine Learning Research},
   volume  = {15},
   year    = {2014},
-  type    = {Journal Article}
 }
 
 @article{RN150,
@@ -76,16 +75,17 @@
   doi     = {10.7717/peerj-cs.1516},
   url     = {https://www.ncbi.nlm.nih.gov/pubmed/37705656},
   year    = {2023},
-  type    = {Journal Article}
 }
 
 @book{RN162,
   author  = {Kruschke, John K.},
   title   = {Doing Bayesian Data Analysis},
-  edition = {1st Edition},
+  edition = {Second Edition},
+  publisher={Academic Press},
   isbn    = {9780123814852},
-  year    = {2010},
-  type    = {Book}
+  year    = {2015},
+  type    = {Book},
+  doi     = {http://dx.doi.org/10.1016/B978-0-12-405888-0.00001-5}
 }
 
 @article{RN144,
@@ -95,7 +95,7 @@
   volume  = {12},
   pages   = {171-178},
   year    = {1985},
-  type    = {Journal Article}
+  url     = {http://www.jstor.org/stable/4615982},
 }
 
 
@@ -106,7 +106,7 @@
   volume  = {7},
   number  = {4},
   year    = {1992},
-  type    = {Journal Article}
+  doi     = {10.1214/ss/1177011136}
 }
 
 @article{RN153,
@@ -121,7 +121,6 @@
   doi     = {10.1021/ac60319a011},
   url     = {https://www.ncbi.nlm.nih.gov/pubmed/22324584},
   year    = {1972},
-  type    = {Journal Article}
 }
 
 @article{RN149,
@@ -136,7 +135,6 @@
   doi     = {10.1002/biot.201700141},
   url     = {https://www.ncbi.nlm.nih.gov/pubmed/29283217},
   year    = {2018},
-  type    = {Journal Article}
 }
 
 @article{RN148,
@@ -151,7 +149,6 @@
   doi     = {10.1002/1097-0290(20010205)72:3<346::aid-bit12>3.0.co;2-x},
   url     = {https://www.ncbi.nlm.nih.gov/pubmed/11135205},
   year    = {2001},
-  type    = {Journal Article}
 }
 
 @article{RN145,
@@ -165,7 +162,6 @@
              1573-1375},
   doi     = {10.1007/s11222-016-9696-4},
   year    = {2016},
-  type    = {Journal Article}
 }
 
 @article{RN146,
@@ -175,7 +171,6 @@
   volume  = {11},
   pages   = {3571-3594},
   year    = {2010},
-  type    = {Journal Article}
 }
 
 @article{RN147,
@@ -187,7 +182,6 @@
   issn    = {2475-9066},
   doi     = {10.21105/joss.01143},
   year    = {2019},
-  type    = {Journal Article}
 }
 
 @article{harris2020array,
@@ -220,6 +214,7 @@
   number={6},
   pages={2622--2630},
   year={2012},
+	doi={https://doi.org/10.1021/ac202124t},
   publisher={ACS Publications}
 }
 
@@ -231,6 +226,7 @@
   number={14},
   pages={7345--7355},
   year={2015},
+	doi={https://doi.org/10.1021/acs.analchem.5b01521},
[…]
