### 2024-11-25 9315947
unite conclusion and summary, then remove conclusion
