### 2024-11-25 6732e12
test protecting cases
