+++ paper/literature.bib
@@ -153,7 +153,7 @@
 
 @article{RN145,
   author  = {Vehtari, Aki and Gelman, Andrew and Gabry, Jonah},
-  title   = {Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC},
+  title   = {{Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC}},
   journal = {Statistics and Computing},
   volume  = {27},
   number  = {5},
