+++ paper/literature.bib
@@ -46,7 +46,7 @@
 
 @misc{matplotlibzenodo,
   author    = {{The Matplotlib Development Team}},
-  title     = {Matplotlib: Visualization with Python},
+  title     = {{Matplotlib: Visualization with Python}},
   keywords  = {software},
   month     = may,
   year      = 2024,
@@ -58,7 +58,7 @@
 
 @article{RN173,
   author  = {Hoffmann, Matthew D. and Gelman, Andrew},
-  title   = {The No-U-Turn Sampler: Adaptively Setting Path Lengths in Hamiltonian Monte Carlo},
+  title   = {{The No-U-Turn Sampler: Adaptively Setting Path Lengths in Hamiltonian Monte Carlo}},
   journal = {Journal of Machine Learning Research},
   volume  = {15},
   year    = {2014},
@@ -66,7 +66,7 @@
 
 @article{RN150,
   author  = {Abril-Pla, O. and Andreani, V. and Carroll, C. and Dong, L. and Fonnesbeck, C. J. and Kochurov, M. and Kumar, R. and Lao, J. and Luhmann, C. C. and Martin, O. A. and Osthege, M. and Vieira, R. and Wiecki, T. and Zinkov, R.},
-  title   = {{PyMC}: a modern, and comprehensive probabilistic programming framework in Python},
+  title   = {{PyMC}: a modern, and comprehensive probabilistic programming framework in {P}ython},
   journal = {PeerJ Comput Sci},
   volume  = {9},
   pages   = {e1516},
@@ -79,7 +79,7 @@
 
 @book{RN162,
   author  = {Kruschke, John K.},
-  title   = {Doing Bayesian Data Analysis},
+  title   = {{Doing Bayesian Data Analysis}},
   edition = {Second Edition},
   publisher={Academic Press},
   isbn    = {9780123814852},
@@ -101,7 +101,7 @@
 
 @article{RN152,
   author  = {Gelman, Andrew and Rubin, Donald B.},
-  title   = {Inference from Iterative Simulation Using Multiple Sequences},
+  title   = {{Inference from Iterative Simulation Using Multiple Sequences}},
   journal = {Statistical Science},
   volume  = {7},
   number  = {4},
@@ -111,7 +111,7 @@
 
 @article{RN153,
   author  = {Grushka, E.},
-  title   = {Characterization of exponentially modified Gaussian peaks in chromatography},
+  title   = {{Characterization of exponentially modified Gaussian peaks in chromatography}},
   journal = {Anal Chem},
   volume  = {44},
   number  = {11},
@@ -125,7 +125,7 @@
 
 @article{RN149,
   author  = {Hemmerich, J. and Noack, S. and Wiechert, W. and Oldiges, M.},
-  title   = {Microbioreactor Systems for Accelerated Bioprocess Development},
+  title   = {{Microbioreactor Systems for Accelerated Bioprocess Development}},
   journal = {Biotechnol J},
   volume  = {13},
   number  = {4},
@@ -166,7 +166,7 @@
 
 @article{RN146,
   author  = {Watanabe, Sumio},
-  title   = {Asymptotic Equivalence of Bayes Cross Validation and Widely Applicable Information Criterion in Singular Learning Theory},
+  title   = {{Asymptotic Equivalence of Bayes Cross Validation and Widely Applicable Information Criterion in Singular Learning Theory}},
   journal = {Journal of machine learning research},
   volume  = {11},
   pages   = {3571-3594},
@@ -175,7 +175,7 @@
 
 @article{RN147,
   author  = {Kumar, Ravin and Carroll, Colin and Hartikainen, Ari and Martin, Osvaldo},
[…]
