+++ paper/literature.bib
@@ -207,7 +207,7 @@
 }
 
 @article{vivo2012bayesian,
-  title={Bayesian approach for peak detection in two-dimensional chromatography},
+  title={{Bayesian approach for peak detection in two-dimensional chromatography}},
   author={Viv{\'o}-Truyols, Gabriel},
   journal={Analytical chemistry},
   volume={84},
@@ -219,7 +219,7 @@
 }
 
 @article{woldegebriel2015probabilistic,
-  title={Probabilistic model for untargeted peak detection in LC--MS using Bayesian statistics},
+  title={{Probabilistic model for untargeted peak detection in LC--MS using Bayesian statistics}},
   author={Woldegebriel, Michael and Viv{\'o}-Truyols, Gabriel},
   journal={Analytical chemistry},
   volume={87},
@@ -231,7 +231,7 @@
 }
 
 @article{briskot2019prediction,
-  title={Prediction uncertainty assessment of chromatography models using Bayesian inference},
+  title={{Prediction uncertainty assessment of chromatography models using Bayesian inference}},
   author={Briskot, Till and St\"{u}ckler, Ferdinand and Wittkopp, Felix and Williams, Christopher and Yang, Jessica and Konrad, Susanne and Doninger, Katharina and Griesbach, Jan and Bennecke, Moritz and Hepbildikler, Stefan and others},
   journal={Journal of Chromatography A},
   volume={1587},
@@ -242,7 +242,7 @@
 }
 
 @article{yamamoto2021uncertainty,
-  title={Uncertainty quantification for chromatography model parameters by Bayesian inference using sequential Monte Carlo method},
+  title={{Uncertainty quantification for chromatography model parameters by Bayesian inference using sequential Monte Carlo method}},
   author={Yamamoto, Yota and Yajima, Tomoyuki and Kawajiri, Yoshiaki},
   journal={Chemical Engineering Research and Design},
   volume={175},
@@ -253,7 +253,7 @@
 }
 
 @article{wiczling2016much,
-  title={How much can we learn from a single chromatographic experiment? A Bayesian perspective},
+  title={{How much can we learn from a single chromatographic experiment? A Bayesian perspective}},
   author={Wiczling, Pawe{\l} and Kaliszan, Roman},
   journal={Analytical chemistry},
   volume={88},
