### 2024-11-25 #7313 @Y0dler
Thanks for the thorough correction. I fixed the typos and introduced more case protection in the bibliography.

Also, I removed more of the workflow description since a detailed one is part of the documentation already.
Since a conclusion generally is just another summary and thus indeed has a sizeable overlap with the - well - summary, as you stated, I united the two sections and removed the conclusion which should help with the word count.

Finally, I added the doi in a bunch of cases but the following citations don't have a doi or I could not find it (including via crossref, in the paper itself, on scholar, and in other papers citing these ones):
  - nutpie (Software, does not have zenodo doi, either)
  - The No-U-Turn Sampler: Adaptively Setting Path Lengths in Hamiltonian Monte Carlo
  - A class of distributions which includes the normal ones (I added the URL, though)
  - Asymptotic Equivalence of Bayes Cross Validation and Widely Applicable Information Criterion in Singular Learning Theory
