### 2024-11-26 #7313 @csoneson
@openjournals/dev - any idea why the reference check does not seem to be working here?
