### 2024-11-26 5581658
Remove http prefix from DOI entries (#43)
