+++ paper/literature.bib
@@ -85,7 +85,7 @@
   isbn    = {9780123814852},
   year    = {2015},
   type    = {Book},
-  doi     = {http://dx.doi.org/10.1016/B978-0-12-405888-0.00001-5}
+  doi     = {10.1016/B978-0-12-405888-0.00001-5}
 }
 
 @article{RN144,
@@ -214,7 +214,7 @@
   number={6},
   pages={2622--2630},
   year={2012},
-	doi={https://doi.org/10.1021/ac202124t},
+	doi={10.1021/ac202124t},
   publisher={ACS Publications}
 }
 
@@ -226,7 +226,7 @@
   number={14},
   pages={7345--7355},
   year={2015},
-	doi={https://doi.org/10.1021/acs.analchem.5b01521},
+	doi={10.1021/acs.analchem.5b01521},
   publisher={ACS Publications}
 }
 
@@ -237,7 +237,7 @@
   volume={1587},
   pages={101--110},
   year={2019},
-	doi={https://doi.org/10.1016/j.chroma.2018.11.076},
+	doi={10.1016/j.chroma.2018.11.076},
   publisher={Elsevier}
 }
 
@@ -248,7 +248,7 @@
   volume={175},
   pages={223--237},
   year={2021},
-	doi={https://doi.org/10.1016/j.cherd.2021.09.003},
+	doi={10.1016/j.cherd.2021.09.003},
   publisher={Elsevier}
 }
 
@@ -260,7 +260,7 @@
   number={1},
   pages={997--1002},
   year={2016},
-	doi={https://doi.org/10.1021/acs.analchem.5b03859},
+	doi={10.1021/acs.analchem.5b03859},
   publisher={ACS Publications}
 }
 
@@ -272,7 +272,7 @@
   number={10},
   pages={1170--1183},
   year={1971},
-	doi={https://doi.org/10.1021/ac60304a011},
+	doi={10.1021/ac60304a011},
   publisher={ACS Publications}
 }
 
@@ -284,6 +284,6 @@
   number={10},
   pages={1184--1195},
   year={1971},
-	doi={https://doi.org/10.1021/ac60304a005},
+	doi={10.1021/ac60304a005},
   publisher={ACS Publications}
 }
