### 2024-11-26 #7313 @michaelosthege
> I don't know the answer, but just looking at the references in the PDF tells me that a lot of them have a DOI field with "https://doi.org/" incorrectly prepended.

thanks for pointing that out! Fixed it in https://github.com/JuBiotech/peak-performance/pull/43
