### 2024-11-27 5fc6476
Update reference to bib file
