### 2024-11-27 #7313 @xuanxu
The syntax used to reference the bib file is incorrect, [this PR](https://github.com/JuBiotech/peak-performance/pull/44) should fix it.
