### 2024-11-30 #7313 @csoneson
Ok, I think we're almost there @Y0dler. A few journal names in the bibliography still needs to be fully spelled out, and I found two more typos:
- line 20: through -> thorough
- line 99: peaks -> peak (or remove 'a')

In addition to this (and as above, you can generate a new proof with `@editorialbot generate pdf`), I will make another post below with the next steps in the process.
