### 2024-12-01 1f8e4b3
spell out all journal names in bibliography
