+++ docs/source/literature.bib
@@ -46,7 +46,7 @@
 
 @misc{matplotlibzenodo,
   author    = {{The Matplotlib Development Team}},
-  title     = {Matplotlib: Visualization with Python},
+  title     = {{Matplotlib: Visualization with Python}},
   keywords  = {software},
   month     = may,
   year      = 2024,
@@ -58,17 +58,16 @@
 
 @article{RN173,
   author  = {Hoffmann, Matthew D. and Gelman, Andrew},
-  title   = {The No-U-Turn Sampler: Adaptively Setting Path Lengths in Hamiltonian Monte Carlo},
+  title   = {{The No-U-Turn Sampler: Adaptively Setting Path Lengths in Hamiltonian Monte Carlo}},
   journal = {Journal of Machine Learning Research},
   volume  = {15},
   year    = {2014},
-  type    = {Journal Article}
 }
 
 @article{RN150,
   author  = {Abril-Pla, O. and Andreani, V. and Carroll, C. and Dong, L. and Fonnesbeck, C. J. and Kochurov, M. and Kumar, R. and Lao, J. and Luhmann, C. C. and Martin, O. A. and Osthege, M. and Vieira, R. and Wiecki, T. and Zinkov, R.},
-  title   = {{PyMC}: a modern, and comprehensive probabilistic programming framework in Python},
-  journal = {PeerJ Comput Sci},
+  title   = {{PyMC}: a modern, and comprehensive probabilistic programming framework in {P}ython},
+  journal = {PeerJ Computer Science},
   volume  = {9},
   pages   = {e1516},
   issn    = {2376-5992 (Electronic)
@@ -76,47 +75,44 @@
   doi     = {10.7717/peerj-cs.1516},
   url     = {https://www.ncbi.nlm.nih.gov/pubmed/37705656},
   year    = {2023},
-  type    = {Journal Article}
 }
 
 @book{RN162,
   author  = {Kruschke, John K.},
-  title   = {Doing Bayesian Data Analysis},
+  title   = {{Doing Bayesian Data Analysis}},
   edition = {Second Edition},
   publisher={Academic Press},
   isbn    = {9780123814852},
   year    = {2015},
   type    = {Book},
-  doi     = {http://dx.doi.org/10.1016/B978-0-12-405888-0.00001-5}
+  doi     = {10.1016/B978-0-12-405888-0.00001-5}
 }
 
 @article{RN144,
   author  = {Azzalini, A.},
   title   = {A class of distributions which includes the normal ones},
-  journal = {Scand. J. Statist.},
+  journal = {Scandinavian Journal of Statistics},
   volume  = {12},
   pages   = {171-178},
   year    = {1985},
-  type    = {Journal Article},
   url     = {http://www.jstor.org/stable/4615982},
 }
 
 
 @article{RN152,
   author  = {Gelman, Andrew and Rubin, Donald B.},
-  title   = {Inference from Iterative Simulation Using Multiple Sequences},
+  title   = {{Inference from Iterative Simulation Using Multiple Sequences}},
   journal = {Statistical Science},
   volume  = {7},
   number  = {4},
   year    = {1992},
-  type    = {Journal Article},
   doi     = {10.1214/ss/1177011136}
 }
 
 @article{RN153,
   author  = {Grushka, E.},
-  title   = {Characterization of exponentially modified Gaussian peaks in chromatography},
-  journal = {Anal Chem},
+  title   = {{Characterization of exponentially modified Gaussian peaks in chromatography}},
+  journal = {Analytical Chemistry},
   volume  = {44},
   number  = {11},
   pages   = {1733-8},
@@ -125,13 +121,12 @@
   doi     = {10.1021/ac60319a011},
   url     = {https://www.ncbi.nlm.nih.gov/pubmed/22324584},
[…]
