+++ paper/literature.bib
@@ -67,7 +67,7 @@
 @article{RN150,
   author  = {Abril-Pla, O. and Andreani, V. and Carroll, C. and Dong, L. and Fonnesbeck, C. J. and Kochurov, M. and Kumar, R. and Lao, J. and Luhmann, C. C. and Martin, O. A. and Osthege, M. and Vieira, R. and Wiecki, T. and Zinkov, R.},
   title   = {{PyMC}: a modern, and comprehensive probabilistic programming framework in {P}ython},
-  journal = {PeerJ Comput Sci},
+  journal = {PeerJ Computer Science},
   volume  = {9},
   pages   = {e1516},
   issn    = {2376-5992 (Electronic)
@@ -91,7 +91,7 @@
 @article{RN144,
   author  = {Azzalini, A.},
   title   = {A class of distributions which includes the normal ones},
-  journal = {Scand. J. Statist.},
+  journal = {Scandinavian Journal of Statistics},
   volume  = {12},
   pages   = {171-178},
   year    = {1985},
@@ -112,7 +112,7 @@
 @article{RN153,
   author  = {Grushka, E.},
   title   = {{Characterization of exponentially modified Gaussian peaks in chromatography}},
-  journal = {Anal Chem},
+  journal = {Analytical Chemistry},
   volume  = {44},
   number  = {11},
   pages   = {1733-8},
@@ -126,7 +126,7 @@
 @article{RN149,
   author  = {Hemmerich, J. and Noack, S. and Wiechert, W. and Oldiges, M.},
   title   = {{Microbioreactor Systems for Accelerated Bioprocess Development}},
-  journal = {Biotechnol J},
+  journal = {Biotechnology Journal},
   volume  = {13},
   number  = {4},
   pages   = {e1700141},
@@ -140,7 +140,7 @@
 @article{RN148,
   author  = {Kostov, Y. and Harms, P. and Randers-Eichhorn, L. and Rao, G.},
   title   = {Low-cost microbioreactor for high-throughput bioprocessing},
-  journal = {Biotechnol Bioeng},
+  journal = {Biotechnology and Bioengineering},
   volume  = {72},
   number  = {3},
   pages   = {346-52},
@@ -167,7 +167,7 @@
 @article{RN146,
   author  = {Watanabe, Sumio},
   title   = {{Asymptotic Equivalence of Bayes Cross Validation and Widely Applicable Information Criterion in Singular Learning Theory}},
-  journal = {Journal of machine learning research},
+  journal = {Journal of Machine Learning Research},
   volume  = {11},
   pages   = {3571-3594},
   year    = {2010},
@@ -209,7 +209,7 @@
 @article{vivo2012bayesian,
   title={{Bayesian approach for peak detection in two-dimensional chromatography}},
   author={Viv{\'o}-Truyols, Gabriel},
-  journal={Analytical chemistry},
+  journal={Analytical Chemistry},
   volume={84},
   number={6},
   pages={2622--2630},
@@ -221,7 +221,7 @@
 @article{woldegebriel2015probabilistic,
   title={{Probabilistic model for untargeted peak detection in LC--MS using Bayesian statistics}},
   author={Woldegebriel, Michael and Viv{\'o}-Truyols, Gabriel},
-  journal={Analytical chemistry},
+  journal={Analytical Chemistry},
   volume={87},
   number={14},
   pages={7345--7355},
@@ -255,7 +255,7 @@
 @article{wiczling2016much,
   title={{How much can we learn from a single chromatographic experiment? A Bayesian perspective}},
   author={Wiczling, Pawe{\l} and Kaliszan, Roman},
-  journal={Analytical chemistry},
+  journal={Analytical Chemistry},
   volume={88},
   number={1},
   pages={997--1002},
