+++ docs/source/literature.bib
@@ -3,7 +3,8 @@
   keywords = {Software},
   license  = {MIT},
   title    = {{nutpie}},
-  url      = {https://github.com/pymc-devs/nutpie}
+  url      = {https://github.com/pymc-devs/nutpie},
+  year    = {2022},
 }
 
 @article{scipy,
