### 2024-12-02 #7313 @Y0dler
I checked the manuscript again in detail and found some last minor corrections (e.g. the `nutpie` citation was missing a date, `nutpie` was written in regular font once etc.).

Regarding our final tasks, I'll update this checklist:
