### 2024-12-02 #7313 @Y0dler
@csoneson All our tasks should be complete. If you set the version number and the doi, we're good to go from our side 👍
