### 2024-12-02 #7313 @Y0dler
I added the paper title as an additional title to the linked zenodo version. Is that okay or does it have to be the main title?
