### 2024-12-05 #7313 @Y0dler
@csoneson I aligned the order of the authors with the paper and changed the license to "GNU Affero General Public License v3.0 only".
