### 2024-12-05 #7313 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
