### 2024-12-10 #7313 @Kevin-Mattheus-Moerman
- [ ] The following words are European English: `focussed, analysed`, while the following are American English: `analyze, optimization, utilizing, idealized, realized, visualizations, visualization, conceptualized`. We suggest you choose one form and consistently use it.
- [ ] On the Zenodo archive I would recommend removing the statements in brackets for the authors.
They seem unusual and the `Related person` sounds like `Family member` e.g. a brother, to me. Do you mean to say `contributor`? Removing these is not a requirement, I am merely stating that they appear unnecessary and are perhaps confusing/wrong.
