### 2024-12-17 #7313 @Adafede
> And a special thank you to the reviewers: @Adafede, @lazear !!

Was a great experience! Any news on https://github.com/openjournals/joss/issues/813?
