### 2024-12-18 #7313 @michaelosthege
Thank you all for the thorough reviews and your patience with us learning how to do our first JOSS paper :)

And thanks for the reminder to update our ORCID entries!
